Add files via upload
The first project that I made using Python! A simple GUI that I created using PySimpleGUI to automate some of the manual tasks (manual writing/ data logging) for my company. I made use of PySimpleGUI to create an interface that obtains input data from the user, and combine it with the pandas library to append the acquired information into an excel sheet to ease tracking and analysis for our engineers on some of our work related equipment.