Data Entry GUI (Graphical User Interface)
My first ever Python Project! This is a a simple GUI that I have created using PySimpleGUI and the Pandas library to automate some of the more manual tasks (Manual written logs/ handwritten reports ) for my company. By using the GUI, users are required to input the necessary information as prompted. Then i made use of the Pandas library to append and store the obtained information into an external excel sheet for the engineers to keep track/analyse the input data much more efficiently.