Data Entry GUI with PySImpleGUI and Pandas
My first ever Python Project! A simple GUI I made using PySimpleGUI in combination with the Pandas library to automate some of the manual tasks (handwritten reports/ manual data logging on forms) for my company. Using PySimpleGUI, the user is required to input the necessary information as prompted. Then, using the pandas library, the obtained information is appended and stored in an external excel sheet.